+++ src/conv-ccs/ccs-builtins.h
@@ -27,13 +27,13 @@ class PUP_fmt : public PUP::wrap_er {
         typeCode_sync=11, // synchronization code
         typeCode_pointer=12 // 32 or 64 bit pointer, depending on the machine architecture
     } typeCode_t;
-    void fieldHeader(typeCode_t typeCode,int nItems);
+    void fieldHeader(typeCode_t typeCode,size_t nItems);
 public:
     PUP_fmt(PUP::er &parent_) 
         :PUP::wrap_er(parent_,PUP::er::IS_COMMENTS) {}
     
     virtual void comment(const char *message);
     virtual void synchronize(unsigned int m);
-    virtual void bytes(void *p,int n,size_t itemSize,PUP::dataType t);
+    virtual void bytes(void *p,size_t n,size_t itemSize,PUP::dataType t);
 };
