+++ src/util/pup_stl.h
@@ -42,7 +42,7 @@ namespace PUP {
   inline void operator|(er &p,typename std::basic_string<charType> &v);
   inline void operator|(er &p,std::string &v);
   template <class container>
-  inline int PUP_stl_container_size(er &p,container &c);
+  inline size_t PUP_stl_container_size(er &p,container &c);
   template <class container, class dtype>
   inline void PUP_stl_container_items(er &p,container &c);
   template <> inline void PUP_stl_container_items<std::vector<bool>,bool>(er &p, std::vector<bool> &c);
@@ -88,7 +88,7 @@ namespace PUP {
   template <class charType> 
   inline void operator|(er &p,typename std::basic_string<charType> &v)
   {
-    int nChar=v.length();
+    size_t nChar=v.length();
     p|nChar;
     if (p.isUnpacking()) { //Unpack to temporary buffer
       charType *buf=new charType[nChar];
@@ -104,7 +104,7 @@ namespace PUP {
   inline void operator|(er &p,std::string &v)
   {
     p.syncComment(sync_begin_object,"std::string");
-    int nChar=v.length();
+    size_t nChar=v.length();
     p|nChar;
     if (p.isUnpacking()) { //Unpack to temporary buffer
       char *buf=new char[nChar];
@@ -123,8 +123,8 @@ namespace PUP {
 
   //Impl. util: pup the length of a container
   template <class container>
-  inline int PUP_stl_container_size(er &p,container &c) {
-    int nElem=c.size();
+  inline size_t PUP_stl_container_size(er &p,container &c) {
+    size_t nElem=c.size();
     p|nElem;
     return nElem; 
   }
@@ -157,11 +157,11 @@ namespace PUP {
   template <class container,class dtype>
   inline void PUP_stl_container(er &p,container &c) {
     p.syncComment(sync_begin_array);
-    int nElem=PUP_stl_container_size(p,c);
+    size_t nElem=PUP_stl_container_size(p,c);
     if (p.isUnpacking()) 
       { //Unpacking: Extract each element and push_back:
 	c.resize(0);
-	for (int i=0;i<nElem;i++) {
+	for (size_t i=0;i<nElem;i++) {
 	  p.syncComment(sync_item);
 	  dtype n;
 	  p|n;
@@ -176,10 +176,10 @@ namespace PUP {
   template <class container,class dtype>
   inline void PUP_stl_map(er &p,container &c) {
     p.syncComment(sync_begin_list);
-    int nElem=PUP_stl_container_size(p,c);
+    size_t nElem=PUP_stl_container_size(p,c);
     if (p.isUnpacking()) 
       { //Unpacking: Extract each element and insert:
-	for (int i=0;i<nElem;i++) {
+	for (size_t i=0;i<nElem;i++) {
 	  dtype n;
 	  p|n;
 	  c.insert(n);
